pyproject.toml
[project]
name = "building-an-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
